type changes

+++ src/base/definitions.jl
@@ -86,7 +86,7 @@ const ω_ref_index = 2
 const P_ref_index = 3
 const Q_ref_index = 4
 
-const MAPPING_DICT = Dict{String, Dict{Symbol, Int64}}
+const MAPPING_DICT = Dict{String, Dict{Symbol, Int}}
 
 const LOCAL_STATE_MAPPING = "local_state_mapping"
 const INPUT_PORT_MAPPING = "input_port_mapping"
